feat: restructuring codes!
