--- dags/hivemind_etl_helpers/src/db/telegram/__init__.py
@@ -1,1 +1,0 @@
-from .tc_chats import TelegramChats
